+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import {Navbar, NavbarMenu, Home, About, Footer} from './Main';
+import {Navbar, NavbarMenu, Home, About, Skills, Footer} from './Main';
 
 const main = ReactDOM.createRoot(document.getElementById('main'));
 main.render(
@@ -10,6 +10,7 @@ main.render(
     <NavbarMenu />
     <Home />
     <About />
+    <Skills />
     <Footer />
   </React.StrictMode>
 );